Added HTTP statuses

+++ src/main/java/com.intexsoft/servlets/BookInCart.java
@@ -36,39 +36,48 @@ public class BookInCart extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String stringUUId = request.getParameter("uuid");
         PrintWriter pw = response.getWriter();
-        try {
-            UUID uuid = UUID.fromString(stringUUId);
-            List<CartHasBook> list = cartHasBookDAO.find(new FindCartHasBookRequest().setCartId(uuid));
-            CartHasBookDTO[] cartHasBookDTOarr = new CartHasBookDTO[list.size()];
+        if (connection == null) {
             pw.println("<html>");
-            int cartHasBookCounter = 0;
-            for (CartHasBook cur : list) {
-                CartHasBookDTO cartHasBookDTO = new CartHasBookDTO();
-                cartHasBookDTO.setCartId(cur.getCartId().toString());
-                cartHasBookDTO.setBookId(cur.getBookId().toString());
-                cartHasBookDTO.setBookCount(cur.getBookCount());
-                cartHasBookDTOarr[cartHasBookCounter] = cartHasBookDTO;
-                pw.println("<h1> bookId = " + cur.getBookId() + "</h1>");
-                Book book = bookDAO.find(new FindBookRequest().setBookId(cur.getBookId())).get(0);
-                pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
-                pw.println("<h1> author = " + book.getAuthor() + "</h1>");
-                pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
-                pw.println("<h1> bookCount = " + cur.getBookCount() + "</h1>");
-                pw.println("<h1> -----------------------</h1>");
-                cartHasBookCounter++;
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
+            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            try {
+                UUID uuid = UUID.fromString(stringUUId);
+                List<CartHasBook> list = cartHasBookDAO.find(new FindCartHasBookRequest().setCartId(uuid));
+                CartHasBookDTO[] cartHasBookDTOarr = new CartHasBookDTO[list.size()];
+                pw.println("<html>");
+                int cartHasBookCounter = 0;
+                for (CartHasBook cur : list) {
+                    CartHasBookDTO cartHasBookDTO = new CartHasBookDTO();
+                    cartHasBookDTO.setCartId(cur.getCartId().toString());
+                    cartHasBookDTO.setBookId(cur.getBookId().toString());
+                    cartHasBookDTO.setBookCount(cur.getBookCount());
+                    cartHasBookDTOarr[cartHasBookCounter] = cartHasBookDTO;
+                    pw.println("<h1> bookId = " + cur.getBookId() + "</h1>");
+                    Book book = bookDAO.find(new FindBookRequest().setBookId(cur.getBookId())).get(0);
+                    pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
+                    pw.println("<h1> author = " + book.getAuthor() + "</h1>");
+                    pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
+                    pw.println("<h1> bookCount = " + cur.getBookCount() + "</h1>");
+                    pw.println("<h1> -----------------------</h1>");
+                    cartHasBookCounter++;
+                }
+                pw.println("<h1> As JSON= " + jsonSerializer.serialize(cartHasBookDTOarr) + "</h1>");
+                pw.println("</html>");
+            } catch (IllegalArgumentException |NullPointerException e) {
+                pw.println("<html>");
+                pw.println("<h1> Invalid UUID </h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            } catch (IndexOutOfBoundsException e) {
+                pw.println("<html>");
+                pw.println("<h1> No such cart an book UUIDs </h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(404);
+                pw.println("</html>");
             }
-            pw.println("<h1> As JSON= " + jsonSerializer.serialize(cartHasBookDTOarr) + "</h1>");
-            pw.println("</html>");
-        } catch (IllegalArgumentException e) {
-            pw.println("<html>");
-            pw.println("<h1> Invalid UUID </h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
-        } catch (IndexOutOfBoundsException e) {
-            pw.println("<html>");
-            pw.println("<h1> No such cart an book UUIDs </h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
         }
     }
 
@@ -76,58 +85,73 @@ public class BookInCart extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            CartHasBookDTO cartHasBookDTO = mapper.map(jsonDeserializer.parseValue(), CartHasBookDTO.class);
-            cartHasBookDAO.createRow(UUID.fromString(cartHasBookDTO.getCartId()), UUID.fromString(cartHasBookDTO.getBookId()), cartHasBookDTO.getBookCount());
+        if (connection == null) {
             pw.println("<html>");
-            pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
-            pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
-            pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                CartHasBookDTO cartHasBookDTO = mapper.map(jsonDeserializer.parseValue(), CartHasBookDTO.class);
+                cartHasBookDAO.createRow(UUID.fromString(cartHasBookDTO.getCartId()), UUID.fromString(cartHasBookDTO.getBookId()), cartHasBookDTO.getBookCount());
+                pw.println("<html>");
+                pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
+                pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
+                pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+                pw.println("</html>");
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
-
 
     @Override
     protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            CartHasBookDTO cartHasBookDTO = mapper.map(jsonDeserializer.parseValue(), CartHasBookDTO.class);
-            cartHasBookDAO.update(new UpdateCartHasBookRequest()
-                            .setCartId(UUID.fromString(cartHasBookDTO.getCartId()))
-                            .setBookId(UUID.fromString(cartHasBookDTO.getBookId()))
-                            .setBookCount(cartHasBookDTO.getBookCount())
-                    , new FindCartHasBookRequest()
-                            .setCartId(UUID.fromString(cartHasBookDTO.getCartId()))
-                            .setBookId(UUID.fromString(cartHasBookDTO.getBookId())));
+        if (connection == null) {
             pw.println("<html>");
-            pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
-            pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
-            pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                CartHasBookDTO cartHasBookDTO = mapper.map(jsonDeserializer.parseValue(), CartHasBookDTO.class);
+                cartHasBookDAO.update(new UpdateCartHasBookRequest()
+                                .setCartId(UUID.fromString(cartHasBookDTO.getCartId()))
+                                .setBookId(UUID.fromString(cartHasBookDTO.getBookId()))
+                                .setBookCount(cartHasBookDTO.getBookCount())
+                        , new FindCartHasBookRequest()
+                                .setCartId(UUID.fromString(cartHasBookDTO.getCartId()))
+                                .setBookId(UUID.fromString(cartHasBookDTO.getBookId())));
+                pw.println("<html>");
+                pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
+                pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
+                pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+                pw.println("</html>");
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
 
@@ -135,34 +159,42 @@ public class BookInCart extends HttpServlet {
     protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            DeleteBookFromCartDTO deleteBookFromCartDTO = mapper.map(jsonDeserializer.parseValue(), DeleteBookFromCartDTO.class);
-            CartHasBook cartHasBook = cartHasBookDAO
-                    .find(new FindCartHasBookRequest()
-                            .setCartId(UUID.fromString(deleteBookFromCartDTO.getCartId()))
-                            .setBookId(UUID.fromString(deleteBookFromCartDTO.getBookId())))
-                    .get(0);
-            CartHasBookDTO cartHasBookDTO = new CartHasBookDTO();
-            cartHasBookDTO.setCartId(cartHasBook.getCartId().toString());
-            cartHasBookDTO.setBookId(cartHasBook.getBookId().toString());
-            cartHasBookDTO.setBookCount(cartHasBook.getBookCount());
+        if (connection == null) {
             pw.println("<html>");
-            pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
-            pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
-            pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-            cartHasBookDAO.delete(new FindCartHasBookRequest().setCartId(cartHasBook.getCartId()).setBookId(cartHasBook.getBookId()));
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                DeleteBookFromCartDTO deleteBookFromCartDTO = mapper.map(jsonDeserializer.parseValue(), DeleteBookFromCartDTO.class);
+                CartHasBook cartHasBook = cartHasBookDAO
+                        .find(new FindCartHasBookRequest()
+                                .setCartId(UUID.fromString(deleteBookFromCartDTO.getCartId()))
+                                .setBookId(UUID.fromString(deleteBookFromCartDTO.getBookId())))
+                        .get(0);
+                CartHasBookDTO cartHasBookDTO = new CartHasBookDTO();
+                cartHasBookDTO.setCartId(cartHasBook.getCartId().toString());
+                cartHasBookDTO.setBookId(cartHasBook.getBookId().toString());
+                cartHasBookDTO.setBookCount(cartHasBook.getBookCount());
+                pw.println("<html>");
+                pw.println("<h1> cartId = " + cartHasBookDTO.getCartId() + "</h1>");
+                pw.println("<h1> bookId = " + cartHasBookDTO.getBookId() + "</h1>");
+                pw.println("<h1> bookCount = " + cartHasBookDTO.getBookCount() + "</h1>");
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(cartHasBookDTO) + "</h1>");
+                pw.println("</html>");
+                cartHasBookDAO.delete(new FindCartHasBookRequest().setCartId(cartHasBook.getCartId()).setBookId(cartHasBook.getBookId()));
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
 }
