+++ src/main/java/com.intexsoft/servlets/BookServlet.java
@@ -21,6 +21,7 @@ import java.util.UUID;
 
 @WebServlet(name = "BookServlet", value = "/bookServlet")
 public class BookServlet extends HttpServlet {
+
     ConnectToDb connectToDb = new ConnectToDb();
     Connection connection = connectToDb.connect_to_db("bookshop", "postgres", "postgrespw");
     BookDAO bookDAO = new BookDAO(connection);
@@ -28,39 +29,50 @@ public class BookServlet extends HttpServlet {
     JsonSerializer jsonSerializer = new JsonSerializer();
     Mapper mapper = new Mapper();
 
+
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String stringUUId = request.getParameter("uuid");
-        PrintWriter pw = response.getWriter();
-        try {
-            UUID uuid = UUID.fromString(stringUUId);
-            Book book = bookDAO.find(new FindBookRequest().setBookId(uuid)).get(0);
-            BookDTO bookDTO = new BookDTO();
-            bookDTO.setBookId(book.getBookId().toString());
-            bookDTO.setBookname(book.getBookname());
-            bookDTO.setAuthor(book.getAuthor());
-            bookDTO.setCostInByn(book.getCostInByn());
-            bookDTO.setCountInStock(book.getCountInStock());
 
-            String json = jsonSerializer.serialize(bookDTO);
+        PrintWriter pw = response.getWriter();
+        if (connection == null) {
             pw.println("<html>");
-            pw.println("<h1> book_id = " + book.getBookId() + "</h1>");
-            pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
-            pw.println("<h1> author = " + book.getAuthor() + "</h1>");
-            pw.println("<h1> cost_in_byn = " + book.getCostInByn() + "</h1>");
-            pw.println("<h1> count_in_stock = " + book.getCountInStock() + "</h1>");
-            pw.println("<h1> As JSON = " + json + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-        } catch (IllegalArgumentException e) {
-            pw.println("<html>");
-            pw.println("<h1> Invalid UUID </h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
-        } catch (IndexOutOfBoundsException e) {
-            pw.println("<html>");
-            pw.println("<h1> No such bookId UUID </h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            try {
+                UUID uuid = UUID.fromString(stringUUId);
+                Book book = bookDAO.find(new FindBookRequest().setBookId(uuid)).get(0);
+                BookDTO bookDTO = new BookDTO();
+                bookDTO.setBookId(book.getBookId().toString());
+                bookDTO.setBookname(book.getBookname());
+                bookDTO.setAuthor(book.getAuthor());
+                bookDTO.setCostInByn(book.getCostInByn());
+                bookDTO.setCountInStock(book.getCountInStock());
+
+                String json = jsonSerializer.serialize(bookDTO);
+                pw.println("<html>");
+                pw.println("<h1> book_id = " + book.getBookId() + "</h1>");
+                pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
+                pw.println("<h1> author = " + book.getAuthor() + "</h1>");
+                pw.println("<h1> cost_in_byn = " + book.getCostInByn() + "</h1>");
+                pw.println("<h1> count_in_stock = " + book.getCountInStock() + "</h1>");
+                pw.println("<h1> As JSON = " + json + "</h1>");
+                pw.println("</html>");
+            } catch (IllegalArgumentException | NullPointerException e) {
+                pw.println("<html>");
+                pw.println("<h1> Invalid UUID </h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            } catch (IndexOutOfBoundsException e) {
+                pw.println("<html>");
+                pw.println("<h1> No such bookId </h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(404);
+                pw.println("</html>");
+            }
         }
     }
 
@@ -69,33 +81,41 @@ public class BookServlet extends HttpServlet {
 
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            CreateBookDTO createBookDTO = mapper.map(jsonDeserializer.parseValue(), CreateBookDTO.class);
+        if (connection == null) {
             pw.println("<html>");
-            Book book = bookDAO.createRow(createBookDTO.getBookname(), createBookDTO.getAuthor(), createBookDTO.getCostInByn(), createBookDTO.getCountInStock());
-            pw.println("<h1> bookid = " + book.getBookId() + "</h1>");
-            pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
-            pw.println("<h1> author = " + book.getAuthor() + "</h1>");
-            pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
-            pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
-            BookDTO bookDTO = new BookDTO();
-            bookDTO.setAuthor(book.getAuthor());
-            bookDTO.setBookname(book.getBookname());
-            bookDTO.setCountInStock(book.getCountInStock());
-            bookDTO.setCostInByn(book.getCostInByn());
-            bookDTO.setBookId(book.getBookId().toString());
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                CreateBookDTO createBookDTO = mapper.map(jsonDeserializer.parseValue(), CreateBookDTO.class);
+                pw.println("<html>");
+                Book book = bookDAO.createRow(createBookDTO.getBookname(), createBookDTO.getAuthor(), createBookDTO.getCostInByn(), createBookDTO.getCountInStock());
+                pw.println("<h1> bookid = " + book.getBookId() + "</h1>");
+                pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
+                pw.println("<h1> author = " + book.getAuthor() + "</h1>");
+                pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
+                pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
+                BookDTO bookDTO = new BookDTO();
+                bookDTO.setAuthor(book.getAuthor());
+                bookDTO.setBookname(book.getBookname());
+                bookDTO.setCountInStock(book.getCountInStock());
+                bookDTO.setCostInByn(book.getCostInByn());
+                bookDTO.setBookId(book.getBookId().toString());
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+                pw.println("</html>");
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
 
@@ -103,33 +123,41 @@ public class BookServlet extends HttpServlet {
     protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            BookDTO bookDTO = mapper.map(jsonDeserializer.parseValue(), BookDTO.class);
-            Book book = new Book();
+        if (connection == null) {
             pw.println("<html>");
-            book.setBookId(UUID.fromString(bookDTO.getBookId()));
-            pw.println("<h1> bookId = " + book.getBookId() + "</h1>");
-            book.setBookname(bookDTO.getBookname());
-            pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
-            book.setAuthor(bookDTO.getAuthor());
-            pw.println("<h1> author = " + book.getAuthor() + "</h1>");
-            book.setCostInByn(bookDTO.getCostInByn());
-            pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
-            book.setCountInStock(bookDTO.getCountInStock());
-            pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
-            bookDAO.updateBook(book);
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                BookDTO bookDTO = mapper.map(jsonDeserializer.parseValue(), BookDTO.class);
+                Book book = new Book();
+                pw.println("<html>");
+                book.setBookId(UUID.fromString(bookDTO.getBookId()));
+                pw.println("<h1> bookId = " + book.getBookId() + "</h1>");
+                book.setBookname(bookDTO.getBookname());
+                pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
+                book.setAuthor(bookDTO.getAuthor());
+                pw.println("<h1> author = " + book.getAuthor() + "</h1>");
+                book.setCostInByn(bookDTO.getCostInByn());
+                pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
+                book.setCountInStock(bookDTO.getCountInStock());
+                pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
+                bookDAO.updateBook(book);
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+                pw.println("</html>");
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
 
@@ -137,35 +165,43 @@ public class BookServlet extends HttpServlet {
     protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         ServletInputStream servletInputStream = request.getInputStream();
         PrintWriter pw = response.getWriter();
-        StringBuilder sb = new StringBuilder();
-        while (!servletInputStream.isFinished()) {
-            sb.append(Character.valueOf((char) servletInputStream.read()));
-        }
-        try {
-            JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
-            DeleteDTO deleteDTO = mapper.map(jsonDeserializer.parseValue(), DeleteDTO.class);
-            Book book = bookDAO.find(new FindBookRequest().setBookId(UUID.fromString(deleteDTO.getUuid()))).get(0);
+        if (connection == null) {
             pw.println("<html>");
-            BookDTO bookDTO = new BookDTO();
-            bookDTO.setBookId(book.getBookId().toString());
-            pw.println("<h1> bookId = " + book.getBookId() + "</h1>");
-            bookDTO.setBookname(book.getBookname());
-            pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
-            bookDTO.setAuthor(book.getAuthor());
-            pw.println("<h1> author = " + book.getAuthor() + "</h1>");
-            bookDTO.setCostInByn(book.getCostInByn());
-            pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
-            bookDTO.setCountInStock(book.getCountInStock());
-            pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
-            bookDAO.delete(new FindBookRequest().setBookId(book.getBookId()));
-            pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+            pw.println("<h1> Problem with connection to server's DateBase</h1>");
             pw.println("</html>");
+            response.setStatus(503);
+        } else {
+            StringBuilder sb = new StringBuilder();
+            while (!servletInputStream.isFinished()) {
+                sb.append(Character.valueOf((char) servletInputStream.read()));
+            }
+            try {
+                JsonDeserializer jsonDeserializer = new JsonDeserializer(sb.toString());
+                DeleteDTO deleteDTO = mapper.map(jsonDeserializer.parseValue(), DeleteDTO.class);
+                Book book = bookDAO.find(new FindBookRequest().setBookId(UUID.fromString(deleteDTO.getUuid()))).get(0);
+                pw.println("<html>");
+                BookDTO bookDTO = new BookDTO();
+                bookDTO.setBookId(book.getBookId().toString());
+                pw.println("<h1> bookId = " + book.getBookId() + "</h1>");
+                bookDTO.setBookname(book.getBookname());
+                pw.println("<h1> bookname = " + book.getBookname() + "</h1>");
+                bookDTO.setAuthor(book.getAuthor());
+                pw.println("<h1> author = " + book.getAuthor() + "</h1>");
+                bookDTO.setCostInByn(book.getCostInByn());
+                pw.println("<h1> costInByn = " + book.getCostInByn() + "</h1>");
+                bookDTO.setCountInStock(book.getCountInStock());
+                pw.println("<h1> countInStock = " + book.getCountInStock() + "</h1>");
+                bookDAO.delete(new FindBookRequest().setBookId(book.getBookId()));
+                pw.println("<h1> As JSON = " + jsonSerializer.serialize(bookDTO) + "</h1>");
+                pw.println("</html>");
 
-        } catch (Exception e) {
-            pw.println("<html>");
-            pw.println("<h1> Wrong JSON</h1>");
-            pw.println("<h1>" + e + "</h1>");
-            pw.println("</html>");
+            } catch (Exception e) {
+                pw.println("<html>");
+                pw.println("<h1> Wrong JSON</h1>");
+                pw.println("<h1>" + e + "</h1>");
+                response.setStatus(400);
+                pw.println("</html>");
+            }
         }
     }
 }
